Inlining + warning fix + minor change

+++ GraphicsDevelopmentCoursework/Viewport.cpp
@@ -33,4 +33,4 @@ using namespace std;
 	mat4 Viewport::getProjection() const noexcept { return projection; }
 	mat4 Viewport::getModelview() const noexcept { return modelview; }
 // other
-	void Viewport::call() const { glViewport(posx, posy, (GLsizei)width, (GLsizei)height);  }
+	void Viewport::call() const { glViewport((GLint)posx, (GLint)posy, (GLsizei)width, (GLsizei)height); }
